requirements.txt
gradio==4.44.0
torch==2.2.2
torchvision==0.17.2
pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
